# runtime/internal/openapidocs/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "runtimetypes_AuthFlow": {
        "properties": {
          "extractCookie": {
            "type": "string"
          },
          "extractJsonPath": {
            "type": "string"
          },
          "injectFormat": {
            "type": "string"
          },
          "injectHeader": {
            "type": "string"
          },
          "loginBody": {
            "type": "string"
          },
          "loginMethod": {
            "type": "string"
          },
          "loginUrl": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "runtimetypes_InjectionArg": {
        "properties": {
          "in": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {}
        },
        "type": "object"
      },
...
      "runtimetypes_RemoteTools": {
        "properties": {
          "authFlow": {
            "$ref": "#/components/schemas/runtimetypes_AuthFlow"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "endpointUrl": {
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "id": {
            "type": "string"
          },
          "injectParams": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "insecureSkipVerify": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/runtimetypes_InjectionArg"
          },
          "specUrl": {
            "type": "string"
          },
          "timeoutMs": {
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "taskengine_FunctionTool": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parameters": {}
        },
        "type": "object"
      },
...
      "taskengine_Tool": {
        "properties": {
          "function": {
            "$ref": "#/components/schemas/taskengine_FunctionTool"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "toolsproviderservice_LocalTools": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/taskengine_Tool"
            },
            "type": "array"
          },
          "type": {
            "type": "string"
          },
          "unavailableReason": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP API served by `contenox serve` under /api. Generated from route annotations by tools/openapi-gen.",
    "title": "Contenox Runtime API",
    "version": "v0.36.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tools/local": {
      "get": {
        "operationId": "tools_listLocal",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/toolsproviderservice_LocalTools"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "listLocal returns every locally available tool group (builtin, MCP-backed, and remote) with the tools each currently exposes.",
        "tags": [
          "tools"
        ]
      }
    },
    "/tools/remote": {
      "get": {
        "operationId": "tools_list",
        "parameters": [
          {
            "description": "An optional RFC3339Nano timestamp to fetch the next page of results.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The maximum number of items to return per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/runtimetypes_RemoteTools"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "list returns the registered remote tools providers, paginated by cursor.",
        "tags": [
          "tools"
        ]
      },
      "post": {
        "operationId": "tools_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/runtimetypes_RemoteTools"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_RemoteTools"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "create registers a new remote tools provider and returns it with its assigned ID.",
        "tags": [
          "tools"
        ]
      }
    },
    "/tools/remote/by-name/{name}": {
      "get": {
        "operationId": "tools_getByName",
        "parameters": [
          {
            "description": "The unique name for the remote tool.",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_RemoteTools"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "getByName returns one remote tools provider by its unique name.",
        "tags": [
          "tools"
        ]
      }
    },
    "/tools/remote/{id}": {
      "delete": {
        "operationId": "tools_delete",
        "parameters": [
          {
            "description": "The unique identifier for the remote tool.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "delete removes a remote tools provider by ID.",
        "tags": [
          "tools"
        ]
      },
      "get": {
        "operationId": "tools_get",
        "parameters": [
          {
            "description": "The unique identifier for the remote tool.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_RemoteTools"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "get returns one remote tools provider by ID.",
        "tags": [
          "tools"
        ]
      },
      "put": {
        "operationId": "tools_update",
        "parameters": [
          {
            "description": "The unique identifier for the remote tool.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/runtimetypes_RemoteTools"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_RemoteTools"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "update replaces a remote tools provider's configuration by ID and returns the stored result.",
        "tags": [
          "tools"
        ]
      }
    },
    "/tools/schemas": {
      "get": {
        "operationId": "tools_getSchemas",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "getSchemas returns the OpenAPI schema document for every supported tool, keyed by tool name.",
        "tags": [
          "tools"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}